Move BaseClient to Client Use more modern version of NamedTuples
