+++ src/pybomb/clients/games_client.py
@@ -4,11 +4,11 @@ https://www.giantbomb.com/api/documentation#toc-0-17
 """
 from typing import Any, Dict, List, Optional, Union
 
-from pybomb.clients.base_client import BaseClient, ResponseParam
+from pybomb.clients.client import Client, ResponseParam
 from pybomb.response import Response
 
 
-class GamesClient(BaseClient):
+class GamesClient(Client):
     """Client for the 'games' API resource."""
 
     RESOURCE_NAME = "games"
