+++ tests/test_clients/test_game.py
@@ -24,7 +24,7 @@ class TestGameClient:
     @pytest.fixture
     def mock_requests_get(self) -> MagicMock:
         """Request GET test mock."""
-        with patch("pybomb.clients.base_client.get") as req_mock:
+        with patch("pybomb.clients.client.get") as req_mock:
             yield req_mock
 
     @pytest.fixture
